вынес точки маршрута в отдельный презентер
