+++ src/presenter/presenter.js
@@ -1,8 +1,7 @@
-import { render, replace } from '../framework/render.js';
+import { render } from '../framework/render.js';
 import DrawPoint from '../view/draw_point.js';
-import EditWayPoints from '../view/edit_way_point.js';
 import Filter from '../view/Filter.js';
-import WayPoint from '../view/way_point.js';
+import TripPresenter from './trip-presenter.js';
 export default class Presenter {
   boardComponent = new DrawPoint();
   constructor({container, pointModel}) {
@@ -13,8 +12,7 @@ export default class Presenter {
   init() {
     this.#renderBoardComponent();
     this.#renderFilter();
-    const offers = this.pointModel.getOffers();
-    this.#renderEvents(offers);
+    this.#renderTrip();
   }
 
   #renderBoardComponent() {
@@ -25,48 +23,11 @@ export default class Presenter {
     render(new Filter(), this.container);
   }
 
-  #renderEvents(offers) {
-    offers.forEach((offer) => this.#renderEvent(offer, offers));
+  #renderTrip() {
+    const TripPres = new TripPresenter(this.container, this.pointModel);
+    TripPres.init();
   }
 
-  #renderEvent(offer, offers) {
-    const point = this.pointModel.getPoints();
-    const destination = this.pointModel.getDestination();
-    const onEditClick = () => swithToEditWayPoint();
-    const onFormWayPoint = () => switchToWayPoint();
-    const onEsc = (evt) => {
-      if (evt.key === 'Escape') {
-        evt.preventDefault();
-        switchToWayPoint();
-      }
-    };
-
-    const wayPoint = new WayPoint(
-      point,
-      destination,
-      offer,
-      offers,
-      onFormWayPoint
-    );
-    const eventEditView = new EditWayPoints(
-      offers,
-      offer,
-      destination,
-      point,
-      onEditClick);
-
-    function swithToEditWayPoint() {
-      replace(wayPoint, eventEditView);
-      document.addEventListener('keydown', onEsc);
-    }
-
-    function switchToWayPoint() {
-      replace(eventEditView, wayPoint);
-      document.removeEventListener('keydown', onEsc);
-    }
-
-    render(wayPoint, this.container);
-  }
 }
